api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
qdrant-client==1.9.1
sentence-transformers==2.7.0
google-generativeai==0.7.2
python-dotenv==1.0.1
numpy==1.26.4
